Update README.me File

+++ app/Ecommerce/Backend/Controllers/Admin/Product/Controllers/ProductVariantController.php
@@ -6,6 +6,7 @@ use App\DataTables\ProductVariantDataTable;
 use App\Http\Controllers\Controller;
 use Ecommerce\Backend\Controllers\Admin\Product\Models\Product;
 use Ecommerce\Backend\Controllers\Admin\Product\Models\ProductVariant;
+use Ecommerce\Backend\Controllers\Admin\Product\Models\ProductVariantItem;
 use Illuminate\Http\Request;
 
 class ProductVariantController extends Controller
@@ -92,9 +93,11 @@ class ProductVariantController extends Controller
      */
     public function destroy(string $id)
     {
-
         $varinat = ProductVariant::findOrFail($id);
-
+        $variantItemCheck = ProductVariantItem::where('product_variant_id', $varinat->id)->count();
+        if($variantItemCheck > 0){
+            return response(['status' => 'error', 'message' => 'This variant contain variant items in it delete the variant items first for delete this variant!']);
+        }
         $varinat->delete();
 
         return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
